General linting and filebreakups
