+++ pyAIAgent/game/data.py
@@ -1,269 +1,5 @@
-# helpers.py (GEN 1)
-import struct
-import pathlib
-from PIL import Image, ImageDraw
-from dump import find_path, dump_minimal_map, dump_minimap_map_array
-import time
-import json
-import re
-from tqdm import tqdm
 from enum import IntEnum
 
-DEFAULT_ROM = 'c.gbc'
-
-# Image capture dimensions
-GBA_WIDTH = 240
-GBA_HEIGHT = 160
-GB_WIDTH = 160
-GB_HEIGHT = 144
-BYTES_PER_PIXEL = 4
-MINI_MAP_SIZE = (21,21)  # 21x21 tiles for the minimap, None = Full Map
-
-
-GBA_RASTER_SIZE = GBA_WIDTH * GBA_HEIGHT * BYTES_PER_PIXEL
-GB_RASTER_SIZE = GB_WIDTH * GB_HEIGHT * BYTES_PER_PIXEL
-SIZE_MAP = {
-    GBA_RASTER_SIZE: (GBA_WIDTH, GBA_HEIGHT),
-    GB_RASTER_SIZE: (GB_WIDTH, GB_HEIGHT),
-}
-
-
-def _flush_socket(sock) -> None:
-    """
-    Drain any pending data from sock so that our next recv()
-    only sees the fresh response to the command we send.
-    """
-    # Switch to non-blocking so recv() returns immediately if no data
-    sock.setblocking(False)
-    try:
-        while True:
-            data = sock.recv(4096)
-            if not data:
-                break
-    except (BlockingIOError, OSError):
-        # No more data to read
-        pass
-    finally:
-        # Go back to blocking mode
-        sock.setblocking(True)
-
-
-def capture(sock, filename: str = "latest.png", cell_size: int = 16) -> None:
-    # flush any leftover bytes
-    _flush_socket(sock)
-
-    sock.sendall(b"CAP\n")
-    hdr = sock.recv(4)
-    if len(hdr) < 4:
-        raise RuntimeError("socket closed during CAP header")
-    length = struct.unpack(">I", hdr)[0]
-
-    data = bytearray()
-    while len(data) < length:
-        chunk = sock.recv(length - len(data))
-        if not chunk:
-            raise RuntimeError("socket closed mid-image")
-        data.extend(chunk)
-
-    size = SIZE_MAP.get(length)
-    if size is None:
-        raise RuntimeError(f"unexpected raster size {length} bytes")
-
-    # build image from raw data
-    img = Image.frombytes("RGBA", size, bytes(data), "raw", "ARGB")
-
-    # draw the 16×16 grid
-    draw = ImageDraw.Draw(img)
-    w, h = img.size
-    grid_color = (255, 0, 0, 128)  # semi-transparent red
-
-    for x in range(0, w + 1, cell_size):
-        draw.line(((x, 0), (x, h)), fill=grid_color)
-    for y in range(0, h + 1, cell_size):
-        draw.line(((0, y), (w, y)), fill=grid_color)
-
-    # save
-    path = pathlib.Path(filename)
-    img.save(path)
-
-def readrange(sock, address: str, length: str) -> bytes:
-    _flush_socket(sock)
-    cmd = f"READRANGE {address} {length}\n".encode('utf-8')
-    sock.sendall(cmd)
-    hdr = sock.recv(4)
-    if len(hdr) < 4:
-        raise RuntimeError("socket closed during READRANGE header")
-    size = struct.unpack(">I", hdr)[0]
-    data = bytearray()
-    while len(data) < size:
-        chunk = sock.recv(size - len(data))
-        if not chunk:
-            raise RuntimeError("socket closed mid-dump")
-        data.extend(chunk)
-    return bytes(data)
-
-
-def send_command(sock, cmd: str) -> str:
-    _flush_socket(sock)
-    sock.sendall((cmd.strip() + "\n").encode('utf-8'))
-    data = bytearray()
-    while True:
-        chunk = sock.recv(4096)
-        if not chunk:
-            raise RuntimeError("socket closed before full response")
-        data.extend(chunk)
-        if b"\n" in chunk:
-            break
-    return data.decode('utf-8').rstrip("\n")
-
-def touch_controls_path_find(mapid, currentPos, screenCoords):
-    """
-    Translate the screentouch to worldspace and gets actions to navigate.
-    Player is always at [4,4] ([0,0] is upper left cell)
-    """
-    x = int(screenCoords[0]) - 4
-    y = int(screenCoords[1]) - 4
-    print(f"POS: {int(currentPos[0])},{int(currentPos[1])}, Translated: {x},{y}, Desination: {int(currentPos[0]) + x},{int(currentPos[1]) + y}")
-    destination = [max(int(currentPos[0]) + x, 0), max(int(currentPos[1]) + y, 0)]
-    actions = find_path(DEFAULT_ROM, mapid, currentPos, destination)
-    if(actions == None):
-        return "[PATH BLOCKED OR INVALID UNWALKABLE DESTINATION]\n"
-    return actions # None if there is no valid path
-    
-
-def get_state(sock) -> str:
-    _flush_socket(sock)
-    return send_command(sock, "state")
-
-def get_party_text(sock) -> str:
-    _flush_socket(sock)
-    party = []
-    header = readrange(sock, "0xD163", "8")
-    count = header[0]
-    species_map = get_species_map()
-    for slot in range(count):
-        data_addr = 0xD163 + 0x08 + slot * 44
-        name_addr = 0xD163 + 0x152 + slot * 10
-        d = readrange(sock, hex(data_addr), "44")
-        raw_name = readrange(sock, hex(name_addr), "10")
-        internal_id = header[1 + slot]
-
-        # Now expect 4-tuple: (dex_no, mon_name, type1, type2)
-        dex_no, mon_name, type1, type2 = species_map.get(
-            internal_id,
-            (None, f"ID 0x{internal_id:02X}", None, None)
-        )
-
-        hp_cur = struct.unpack(">H", d[1:3])[0]
-        level = d[0x21]
-        hp_max = struct.unpack(">H", d[0x22:0x24])[0]
-        nickname = decode_pokemon_text(raw_name) or "(no nick)"
-
-        # Build a types string, e.g. "Grass/Poison" or just "Fire"
-        types = type1 if type1 else ""
-        if type2:
-            types += f"/{type2}"
-        
-        mon = {"name": mon_name, "level": level, "type": type1, "hp": hp_cur, "maxHp": hp_max, "nickname": nickname}
-        party.append(mon)
-    return party
-
-
-def get_badges_text(sock) -> str:
-    _flush_socket(sock)
-    raw = readrange(sock, "0xD356", "1")
-    flags = raw[0]
-    names = ["Boulder","Cascade","Thunder","Rainbow","Soul","Marsh","Volcano","Earth"]
-    have = [names[i] for i in range(8) if flags & (1 << i)]
-    return have
-
-
-def get_facing(sock) -> str:
-    _flush_socket(sock)
-    raw = readrange(sock, "0xC109", "1")[0]
-    code = raw & 0xC
-    if code == 0x0:
-        return "down"
-    elif code == 0x4:
-        return "up"
-    elif code == 0x8:
-        return "left"
-    elif code == 0xC:
-        return "right"
-    else:
-        return f"unknown(0x{raw:02X})"
-
-
-def get_location(sock) -> tuple[int, int, int, str] | None:
-    _flush_socket(sock)
-    mid = readrange(sock, "0xD35E", "1")[0]
-    mapName = get_location_name(mid)
-    tile_x = readrange(sock, "0xD362", "1")[0]
-    tile_y = readrange(sock, "0xD361", "1")[0]
-    map_h_blocks = readrange(sock, "0xD368", "1")[0]
-    map_w_blocks = readrange(sock, "0xD369", "1")[0]
-    map_w_tiles = map_w_blocks * 2
-    if map_w_tiles == 0:
-        return None
-    facing = get_facing(sock)
-    return (mid, tile_x, tile_y, facing, mapName)
-
-
-def prep_llm(sock) -> dict:
-    _flush_socket(sock)
-    capture(sock, "latest.png")
-    time.sleep(0.1)
-    _flush_socket(sock)
-    loc = get_location(sock)
-    mid = None
-    mapName = None
-    map2D = ""
-
-    if loc:
-        mid, x, y, facing, mapName = loc
-        dump_minimal_map(DEFAULT_ROM, mid, (x, y), grid_lines=True, crop=MINI_MAP_SIZE).save("minimap.png")
-        map2D = dump_minimap_map_array(DEFAULT_ROM, mid, (x, y), crop=MINI_MAP_SIZE)
-        position = (x, y)
-    else:
-        # no map data or in battle → empty map
-        open("minimap.png", "wb").close()
-        position = None
-        facing = None
-
-    return {
-        "party":   get_party_text(sock),
-        "map_id": mid,
-        "badges":  get_badges_text(sock),
-        "position": position,
-        "facing":  facing,
-        "map_name": mapName,
-        "minimap_2d": map2D
-    }
-
-
-
-def print_battle(sock) -> None:
-    _flush_socket(sock)
-    cur = readrange(sock, hex(0xD057), "1")[0]
-    if cur == 0:
-        print("Not currently in a battle.")
-        return
-    b = readrange(sock, hex(0xD05A), "1")[0]
-    types = {
-        0xF0: "Wild Battle",
-        0xED: "Trainer Battle",
-        0xEA: "Gym Leader Battle",
-        0xF3: "Final Battle",
-        0xF6: "Defeated Trainer",
-        0xF9: "Defeated Wild Pokémon",
-        0xFC: "Defeated Champion/Gym"
-    }
-    label = types.get(b, f"Unknown (0x{b:02X})")
-    print(f"In battle: {label}")
-
-
-
-# ──────────── Internal → Pokédex mapping ────────────
 def get_species_map():
   """
   Returns a dictionary mapping internal Pokémon Red/Blue IDs (hex)
@@ -484,34 +220,6 @@ def decode_pokemon_text(raw_bytes: bytes) -> str:
             out.append('?')
     return ''.join(out)
 
-def parse_optional_fenced_json(text):
-    """
-    Parses JSON from `text`, which may be either:
-      - Enclosed in triple-backtick fences (``` or ```json … ```)
-      - Plain JSON without fences
-
-    Returns:
-        The deserialized Python object.
-
-    Raises:
-        ValueError: if JSON parsing fails or (if fenced) no valid fence is found.
-    """
-    # Try to find a fenced JSON block first
-    fence_pattern = r'```(?:json)?\s*\n(.*?)\n```'
-    m = re.search(fence_pattern, text, flags=re.DOTALL | re.IGNORECASE)
-    if m:
-        json_str = m.group(1)
-    else:
-        # No fence: assume the entire text is JSON
-        json_str = text.strip()
-    
-    try:
-        j = json.loads(json_str)
-        return j
-    except:
-        return {}
-
-# from https://github.com/davidhershey/ClaudePlaysPokemonStarter/blob/main/agent/memory_reader.py
 class MapLocation(IntEnum):
     """Maps location IDs to their names"""
 
@@ -745,83 +453,3 @@ def get_location_name(value: int) -> str | None:
         return MapLocation(value).name
     except ValueError:
         return None
-    
-    
-    
-    
-import os
-import sys
-import shutil
-
-def find_mgba():
-    # os.name == "nt" for window
-    exe_name = "mgba.exe" if os.name == "nt" else "mgba"
-    path = shutil.which(exe_name)
-    if path:
-        print(f"found in {path}")
-        return path
-
-    # fallback for search common install locations
-    search_paths = []
-    
-    if sys.platform.startswith("win"):
-        search_paths = [
-            "C:\\Program Files\\mGBA",
-            "C:\\Program Files (x86)\\mGBA",
-            os.path.expanduser("~\\Downloads\\mGBA")
-        ]
-    elif sys.platform == "darwin":  # macOS
-        search_paths = [
-            "/Applications/mGBA.app/Contents/MacOS",
-            "/usr/local/bin",
-            os.path.expanduser("~/Applications"),
-        ]
-    elif sys.platform.startswith("linux"):
-        search_paths = [
-            "/usr/bin",
-            "/usr/local/bin",
-            os.path.expanduser("~/Downloads"),
-            os.path.expanduser("~/.local/bin"),
-        ]
-
-    for folder in search_paths:
-        full_path = os.path.join(folder, exe_name)
-        if os.path.exists(full_path):
-            return full_path
-    
-    print_dot_interval = 2000 # print dot every 2000 dirs
-    print("Finding mgba file path...", end="", flush=True)
-    
-    # Count total directories first (for accurate progress)
-    
-    dirs = []
-    count = 0
-
-    for root, _, _ in os.walk(os.path.expanduser("~")):
-        dirs.append(root)
-        count += 1
-        if count % print_dot_interval == 0:
-            print(".", end="", flush=True)
-        
-    print()
-    
-
-    for root in tqdm(dirs, desc="Scanning directories"):
-        try:
-            if exe_name in os.listdir(root):
-                return os.path.join(root, exe_name)
-        except (PermissionError, FileNotFoundError):
-            continue
-    print("MGBA not found!")
-    return None
-
-
-def parse_max_loops_fn(value, cli_args):
-    try:                        
-        if value <= 0:
-            return [None, False, "--max_loops value must be a positive integer."]
-            
-        return [value, True, f"Command line argument: --max_loops set to {value}."]
-
-    except ValueError:
-        return [None, False, f"--max_loops expects an integer value, got: {cli_args[i+1]}"]
